DSDEGP-1429: Collapse the V2 ReadGroup back into V1. (#39)
* Back out V2 tests.

* Everything compiles again.

* Scalafmt prefers this.

* A trailing comma crashes scalafmt.

* Patch up the tests -- mostly v2 -=> v1.
