+++ clio-transfer-model/src/main/scala/org/broadinstitute/clio/transfer/model/TransferReadGroupV1QueryInput.scala
@@ -2,11 +2,14 @@ package org.broadinstitute.clio.transfer.model
 
 import java.time.OffsetDateTime
 
-case class TransferReadGroupV1QueryInput(flowcellBarcode: Option[String],
-                                         lane: Option[Int],
-                                         libraryName: Option[String],
-                                         lcSet: Option[String],
-                                         project: Option[String],
-                                         runDateEnd: Option[OffsetDateTime],
-                                         runDateStart: Option[OffsetDateTime],
-                                         sampleAlias: Option[String])
+case class TransferReadGroupV1QueryInput(
+  flowcellBarcode: Option[String],
+  lane: Option[Int],
+  libraryName: Option[String],
+  location: Option[TransferReadGroupLocation],
+  lcSet: Option[String],
+  project: Option[String],
+  runDateEnd: Option[OffsetDateTime],
+  runDateStart: Option[OffsetDateTime],
+  sampleAlias: Option[String]
+)
