+++ clio-transfer-model/src/main/scala/org/broadinstitute/clio/transfer/model/TransferReadGroupV1QueryOutput.scala
@@ -6,6 +6,7 @@ case class TransferReadGroupV1QueryOutput(
   flowcellBarcode: String,
   lane: Int,
   libraryName: String,
+  location: TransferReadGroupLocation,
   analysisType: Option[String],
   baitIntervals: Option[String],
   dataType: Option[String],
