--- clio-transfer-model/src/main/scala/org/broadinstitute/clio/transfer/model/TransferReadGroupV2QueryInput.scala
@@ -1,15 +1,0 @@
-package org.broadinstitute.clio.transfer.model
-
-import java.time.OffsetDateTime
-
-case class TransferReadGroupV2QueryInput(
-  flowcellBarcode: Option[String],
-  lane: Option[Int],
-  libraryName: Option[String],
-  location: Option[TransferReadGroupLocation],
-  lcSet: Option[String],
-  project: Option[String],
-  runDateEnd: Option[OffsetDateTime],
-  runDateStart: Option[OffsetDateTime],
-  sampleAlias: Option[String]
-)
